Add Worker, SharedWorkerState and WorkerPool implementations, refactor MVC accordingly

+++ src/main/java/it/unibo/pcd/assignment1/model/update/Update.java
@@ -3,7 +3,7 @@ package it.unibo.pcd.assignment1.model.update;
 import java.util.Map;
 
 public interface Update {
-    Map<String, Integer> getFrequencies();
+    Map<String, Long> getFrequencies();
 
-    int getProcessedWords();
+    long getProcessedWords();
 }
